+++ myems-api/core/rule.py
@@ -167,8 +167,9 @@ class RuleCollection:
 
         add_row = (" INSERT INTO tbl_rules "
                    "             (name, uuid, category, fdd_code, priority, "
-                   "              channel, expression, message_template, is_enabled) "
-                   " VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s) ")
+                   "              channel, expression, message_template, is_enabled, "
+                   "              is_run_immediately) "
+                   " VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s) ")
         cursor.execute(add_row, (name,
                                  str(uuid.uuid4()),
                                  category,
@@ -177,7 +178,8 @@ class RuleCollection:
                                  channel,
                                  expression,
                                  message_template,
-                                 is_enabled))
+                                 is_enabled,
+                                 False))
         new_id = cursor.lastrowid
         cnx.commit()
         cursor.close()
@@ -403,6 +405,56 @@ class RuleItem:
                                     message_template,
                                     is_enabled,
                                     id_,))
+        cnx.commit()
+
+        cursor.close()
+        cnx.close()
+
+        resp.status = falcon.HTTP_200
+
+
+class RuleRun:
+    @staticmethod
+    def __init__():
+        """Initializes RuleItem"""
+        pass
+
+    @staticmethod
+    def on_options(req, resp, id_):
+        resp.status = falcon.HTTP_200
+
+    @staticmethod
+    @user_logger
+    def on_put(req, resp, id_):
+        """Handles PUT requests"""
+        admin_control(req)
+        try:
+            raw_json = req.stream.read().decode('utf-8')
+        except Exception as ex:
+            raise falcon.HTTPError(status=falcon.HTTP_400,
+                                   title='API.BAD_REQUEST',
+                                   description='API.FAILED_TO_READ_REQUEST_STREAM')
+
+        if not id_.isdigit() or int(id_) <= 0:
+            raise falcon.HTTPError(status=falcon.HTTP_400, title='API.BAD_REQUEST',
+                                   description='API.INVALID_RULE_ID')
+        # ignore the payload
+        new_values = json.loads(raw_json)
+        cnx = mysql.connector.connect(**config.myems_fdd_db)
+        cursor = cnx.cursor()
+
+        cursor.execute(" SELECT id "
+                       " FROM tbl_rules "
+                       " WHERE id = %s ", (id_,))
+        if cursor.fetchone() is None:
+            cursor.close()
+            cnx.close()
+            raise falcon.HTTPError(status=falcon.HTTP_404, title='API.NOT_FOUND',
+                                   description='API.RULE_NOT_FOUND')
+        update_row = (" UPDATE tbl_rules "
+                      " SET is_run_immediately = 1 "
+                      " WHERE id = %s ")
+        cursor.execute(update_row, (id_,))
         cnx.commit()
 
         cursor.close()